+++ app/main/about/[aboutIndex]/1-about.tsx
@@ -1,0 +1,18 @@
+export function AboutContent1() {
+  return (
+    <>
+      <ul className="list-inside list-disc">
+        <li>
+          ダウンロード不要でブラウザーからすぐに遊べる、シンプルでかわいいリズムゲームです。
+        </li>
+        <li>
+          PCだけでなくタブレットやスマートフォンなどでも手軽に遊べるのが特徴で、音楽ゲームにあまり馴染みがない方でも楽しめると思います。
+        </li>
+        <li>
+          さらに、プレイするだけでなく誰でも簡単に譜面を作成でき、それをSNSなどでシェアして遊んでもらうこともできます
+        </li>
+        <li>todo: 概要説明が文章だけでダサいのでなんとかする</li>
+      </ul>
+    </>
+  );
+}
